Conflicts:
	backend/C4A/bin/target/api-1.0-SNAPSHOT/WEB-INF/classes/jdbc.properties
	backend/C4A/src/main/resources/jdbc.properties
	backend/C4A/src/test/resources/unitils.properties

+++ backend/C4A/src/main/java/eu/city4age/dashboard/api/dao/AssessmentDAO.java
@@ -13,7 +13,7 @@ import eu.city4age.dashboard.api.model.TimeInterval;
  */
 public interface AssessmentDAO extends BaseDAO {
 
-	List<TimeInterval> getDiagramDataForUserInRoleId(final Integer crId, final Short dvParentId, final Timestamp start, final Timestamp end);
+	List<TimeInterval> getDiagramDataForUserInRoleId(final Long crId, final Long dvParentId, final Timestamp start, final Timestamp end);
 
 	List<TimeInterval> getLastFiveAssessmentsForDiagram(final Integer crId, final Timestamp start, final Timestamp end);
