+++ backend/C4A/src/main/java/eu/city4age/dashboard/api/dao/DetectionVariableDAO.java
@@ -6,8 +6,8 @@ import eu.city4age.dashboard.api.model.CdDetectionVariable;
 
 public interface DetectionVariableDAO extends BaseDAO {
 	
-	List<String> getAllDetectionVariableNamesForParentId(Short parentId);
+	List<String> getAllDetectionVariableNamesForParentId(Long parentId);
 	
-	List<CdDetectionVariable> getAllDetectionVariableNames(Short parentId);
+	List<CdDetectionVariable> getAllDetectionVariableNames(Long parentId);
 
 }
